requirements.txt
numpy
pdfplumber
PyPDF2
pytesseract
Pillow
tkPDFViewer
